first framework for playlist database saving
the database helper and playlist object have been implemented
successfully, but have not been tested. playlists are stored in the main
musicPlayer class as an arraylist of playlist objects. this will be
where the GUI pulls playlist data from.

in order to actually have the code do anything, we'll need to have the
"save playlist" button get the current displayed music list, save it
into an arraylist of path strings, create a new playlist object from
those, connect to the database, and create the playlist in the database.

we'll also need to query the database to update the main playlist
arraylist every time the software is started.